// packages/openapi/lib/openapi.gen.ts
export type Profile = z.infer<typeof Profile>;
export const Profile = z.object({
  id: z.string(),
  title: z.string(),
  imageUrl: z.union([z.string(), z.null()]),
  isDefault: z.boolean(),
  userId: z.string(),
});

export type ProfileCreate = z.infer<typeof ProfileCreate>;
export const ProfileCreate = z.object({
  title: z.string(),
  imageUrl: z.union([z.string(), z.null(), z.undefined()]).optional(),
});

export type ProfileList = z.infer<typeof ProfileList>;
export const ProfileList = z.object({
  data: z.array(Profile),
  cursor: z.union([z.string(), z.undefined()]).optional(),
});

export type ProfileUpdate = z.infer<typeof ProfileUpdate>;
export const ProfileUpdate = z.object({
  title: z.string().optional(),
  imageUrl: z.union([z.string(), z.null()]).optional(),
});
// ...
export const EndpointByMethod = {
  post: {
    "/auth/register": post_Register,
    "/auth/login": post_Login,
    "/bookmarks": post_CreateBookmark,
    "/collections": post_CreateCollection,
    "/feeds": post_CreateFeed,
    "/feeds/detect": post_DetectFeeds,
    "/feeds/import": post_ImportFeeds,
    "/feeds/export": post_ExportFeeds,
    "/folders": post_CreateFolder,
    "/profiles": post_CreateProfile,
    "/tags": post_CreateTag,
  },
  get: {
    "/auth/@me": get_GetActiveUser,
    "/bookmarks": get_ListBookmarks,
    "/bookmarks/{id}": get_GetBookmark,
    "/collections": get_ListCollections,
    "/collections/{id}": get_GetCollection,
    "/feeds": get_ListFeeds,
    "/feeds/{id}": get_GetFeed,
    "/feedEntries": get_ListFeedEntries,
    "/feedEntries/{id}": get_GetFeedEntry,
    "/folders": get_ListFolders,
    "/folders/{id}": get_GetFolder,
    "/profiles": get_ListProfiles,
    "/profiles/@me": get_GetActiveProfile,
    "/profiles/{id}": get_GetProfile,
    "/tags": get_ListTags,
    "/tags/{id}": get_GetTag,
  },
  patch: {
    "/bookmarks/{id}": patch_UpdateBookmark,
    "/collections/{id}": patch_UpdateCollection,
    "/feeds/{id}": patch_UpdateFeed,
    "/feedEntries/{id}": patch_UpdateFeedEntry,
    "/folders/{id}": patch_UpdateFolder,
    "/profiles/{id}": patch_UpdateProfile,
    "/tags/{id}": patch_UpdateTag,
  },
  delete: {
    "/bookmarks/{id}": delete_DeleteBookmark,
    "/collections/{id}": delete_DeleteCollection,
    "/feeds/{id}": delete_DeleteFeed,
    "/folders/{id}": delete_DeleteFolder,
    "/profiles/{id}": delete_DeleteProfile,
    "/tags/{id}": delete_DeleteTag,
  },
};
export type EndpointByMethod = typeof EndpointByMethod;
// </EndpointByMethod>

// <EndpointByMethod.Shorthands>
export type PostEndpoints = EndpointByMethod["post"];
export type GetEndpoints = EndpointByMethod["get"];
export type PatchEndpoints = EndpointByMethod["patch"];
export type DeleteEndpoints = EndpointByMethod["delete"];
export type AllEndpoints = EndpointByMethod[keyof EndpointByMethod];
// </EndpointByMethod.Shorthands>

// <ApiClientTypes>
export type EndpointParameters = {
  body?: unknown;
  query?: Record<string, unknown>;
  header?: Record<string, unknown>;
  path?: Record<string, unknown>;
};

export type MutationMethod = "post" | "put" | "patch" | "delete";
export type Method = "get" | "head" | MutationMethod;

type RequestFormat = "json" | "form-data" | "form-url" | "binary" | "text";

export type DefaultEndpoint = {
  parameters?: EndpointParameters | undefined;
  response: unknown;
};

export type Endpoint<TConfig extends DefaultEndpoint = DefaultEndpoint> = {
  operationId: string;
  method: Method;
  path: string;
  requestFormat: RequestFormat;
  parameters?: TConfig["parameters"];
  meta: {
    alias: string;
    hasParameters: boolean;
    areParametersRequired: boolean;
  };
  response: TConfig["response"];
};

type Fetcher = (
  method: Method,
  url: string,
  parameters?: EndpointParameters | undefined,
) => Promise<Endpoint["response"]>;

type RequiredKeys<T> = {
  [P in keyof T]-?: undefined extends T[P] ? never : P;
}[keyof T];

type MaybeOptionalArg<T> = RequiredKeys<T> extends never ? [config?: T] : [config: T];

// </ApiClientTypes>

// <ApiClient>
export class ApiClient {
  baseUrl: string = "";

  constructor(public fetcher: Fetcher) {}

  setBaseUrl(baseUrl: string) {
    this.baseUrl = baseUrl;
    return this;
  }

  // <ApiClient.post>
  post<Path extends keyof PostEndpoints, TEndpoint extends PostEndpoints[Path]>(
    path: Path,
    ...params: MaybeOptionalArg<z.infer<TEndpoint["parameters"]>>
  ): Promise<z.infer<TEndpoint["response"]>> {
    return this.fetcher("post", this.baseUrl + path, params[0]) as Promise<z.infer<TEndpoint["response"]>>;
  }
  // </ApiClient.post>

  // <ApiClient.get>
  get<Path extends keyof GetEndpoints, TEndpoint extends GetEndpoints[Path]>(
    path: Path,
    ...params: MaybeOptionalArg<z.infer<TEndpoint["parameters"]>>
  ): Promise<z.infer<TEndpoint["response"]>> {
    return this.fetcher("get", this.baseUrl + path, params[0]) as Promise<z.infer<TEndpoint["response"]>>;
  }
  // </ApiClient.get>

  // <ApiClient.patch>
  patch<Path extends keyof PatchEndpoints, TEndpoint extends PatchEndpoints[Path]>(
    path: Path,
    ...params: MaybeOptionalArg<z.infer<TEndpoint["parameters"]>>
  ): Promise<z.infer<TEndpoint["response"]>> {
    return this.fetcher("patch", this.baseUrl + path, params[0]) as Promise<z.infer<TEndpoint["response"]>>;
  }
  // </ApiClient.patch>

  // <ApiClient.delete>
  delete<Path extends keyof DeleteEndpoints, TEndpoint extends DeleteEndpoints[Path]>(
    path: Path,
    ...params: MaybeOptionalArg<z.infer<TEndpoint["parameters"]>>
  ): Promise<z.infer<TEndpoint["response"]>> {
    return this.fetcher("delete", this.baseUrl + path, params[0]) as Promise<z.infer<TEndpoint["response"]>>;
  }
  // </ApiClient.delete>
}